[업그레이드] 현재골드/필요골드 표시 UI 오류 수정
- BigNumber를 바꾸면서 생긴 오류 → 수정 완료

+++ Assets/01.Scripts/Core/BigNumber.cs
@@ -185,7 +185,8 @@ public struct BigNumber : IComparable<BigNumber>, IEquatable<BigNumber>, IFormat
         if (b.value == 0) throw new DivideByZeroException();
         if (a.value == 0) return new BigNumber(0);
 
-        long newVal = (a.value * ScaleFactor) / b.value;
+        //long newVal = (a.value * ScaleFactor) / b.value;
+        long newVal = (long)(((double)a.value * ScaleFactor) / b.value);
         int newExp = checked(a.exponent - b.exponent);
         return new BigNumber(newVal, newExp);
     }
